requirements.txt
# Core API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# State Management
redis>=5.0.0
motor>=3.3.0

# Financial Data & Processing
yfinance>=0.2.37
finnhub-python>=2.4.19
pandas>=2.2.0
ta>=0.11.0  # Switched to 'ta' for full Python 3.11+ / ARM64 support
requests>=2.31.0

# AI, LLMs & Sentiment (Adjust based on your orchestrator design)
# openai>=1.14.0
# langchain>=0.1.13  # Uncomment if using Langchain
# transformers>=4.38.0  # Uncomment for local HuggingFace sentiment

# Testing & Dev
pytest>=8.1.0
pytest-asyncio>=0.23.5
httpx>=0.27.0
